Remote Acquisition: ImageProvider and RemoteImageProvider to get images locally or elsewhere (#2)
* New files and new module structure

* Modifications to run tests in appropriate sub module

* Modifications to run tests from the root directory

from the command line in the repository directory, it is possible to run the tests like so:

`PYTHONPATH=tests python3 -m unittest discover -s tests`

Callable has been corrected.

* Delete src/pymicroscope/_version.py

We must not track the _version.py file generated automatically

* ImageProviders mechanisms put in place

An imageprovider is anything that can acquire images. It acquires an image and sends it to its "delegate".

* Added a by_uri() method to get pyro objects by URI

* Comments, type hints and better handling of delegates

* Moved DebugRemoteProvider to imageprovider so we can easily test

* Delete _version.py

* ImageProvider can now stream to separate machine and be polled for images

* Better language: client instead of delegate (we can have more than one)

* small improvements to structure