Created a new concept: Configurable objects
We need to get a general -purpose configuration mechanism for many things (Providers, Save, analysis, etc...).

I created ConfigurableProperty which provides all the necessary information to offer a dialog to change the configuration.  Not fully completed, but not bad: we can obtain a configuration_description, then offer a minimal dialog (or a custom one, to do).