Better Provider quit management (#7)
* Create LSMImageProvider.py

* Better name

* Reorganization

* Created a new concept: Configurable objects

We need to get a general -purpose configuration mechanism for many things (Providers, Save, analysis, etc...).

I created ConfigurableProperty which provides all the necessary information to offer a dialog to change the configuration.  Not fully completed, but not bad: we can obtain a configuration_description, then offer a minimal dialog (or a custom one, to do).

* Now not releasing Provider everytime.  Faster but also essential to preserve configuration.

* Completed ImageProvider code to switch from one to the other with the menu several times

* Better available_devices for opencv

Reorganized the UI a bit